+++ ethereum-ra/src/main/java/be/e_contract/ethereum/ra/EthereumConnectionImpl.java
@@ -217,7 +217,7 @@ public class EthereumConnectionImpl implements EthereumConnection {
 
     @Override
     public TransactionReceipt deploy(Class<? extends Contract> contractClass, BigInteger gasPrice,
-            BigInteger gasLimit, Credentials credentials, Integer chainId) throws ResourceException, EthereumException {
+            BigInteger gasLimit, Credentials credentials, Long chainId) throws ResourceException, EthereumException {
         try {
             return this.ethereumManagedConnection.deploy(contractClass, gasPrice, gasLimit, credentials, chainId);
         } catch (EthereumException ex) {
@@ -231,7 +231,7 @@ public class EthereumConnectionImpl implements EthereumConnection {
 
     @Override
     public <T extends Contract> T load(Class<T> contractClass, String contractAddress,
-            Credentials credentials, Integer chainId, BigInteger gasPrice, BigInteger gasLimit) throws ResourceException {
+            Credentials credentials, Long chainId, BigInteger gasPrice, BigInteger gasLimit) throws ResourceException {
         try {
             return this.ethereumManagedConnection.load(contractClass, contractAddress, credentials,
                     chainId, gasPrice, gasLimit);
@@ -242,7 +242,7 @@ public class EthereumConnectionImpl implements EthereumConnection {
     }
 
     @Override
-    public Integer getChainId() throws ResourceException, EthereumException {
+    public Long getChainId() throws ResourceException, EthereumException {
         try {
             return this.ethereumManagedConnection.getChainId();
         } catch (EthereumException ex) {
@@ -256,7 +256,7 @@ public class EthereumConnectionImpl implements EthereumConnection {
 
     @Override
     public String sendTransaction(Credentials credentials, String to, BigInteger value,
-            BigInteger gasPrice, Integer chainId) throws ResourceException, EthereumException {
+            BigInteger gasPrice, Long chainId) throws ResourceException, EthereumException {
         try {
             return this.ethereumManagedConnection.sendTransaction(credentials, to, value, gasPrice, chainId);
         } catch (EthereumException ex) {
